Describe image licensing which is more complicated
License pedantry:

The licensing is, not quite a straightforward as "public domain cc0" because of the images.

I think we should not claim any *additional* restrictions (e.g. creativity in cropping / adding annotations) but the underlying images are not always public domain/cc0  There may be a "fair use" argument for small images, or a map cases an "insubstantial extract", in which case we need to make that argument here maybe.

Some images are clearly OpenStreetMap, and others are software screenshots. We should at least give credits here, but downstream re-users of the content may also need to be aware